Add ArchetypeVolume rotation tests
These are still needing work, they don't pass. This ought to be better
and more succinct with in-game testing being secondary for the cases of
"let's actually see this thing work".

The tests themselves do bootstrap a lot of mocking support objects (due
to the lack of mixins), but overall, this should serve as a valid
testing ground for being able to get archetype volumes almost verified
in their fullest extent (maybe even to storing and reading schematics?).

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>